Split off some logic from lookup-search.jsx.
This file wasn't touched during the view/logic split a few sprints ago, so it
still has a mix of business logic and template code. As it'll be helpful to
reference some of this logic from the homepage, split it off into separate
modules now.
