+++ ui/__tests__/redirects.test.js
@@ -1,0 +1,14 @@
+import { redirectQuery } from '../redirects';
+
+
+describe('redirectQuery()', () => {
+  it('updates an empty query', () => {
+    const result = redirectQuery({ some: 'thing' }, 'myParam', 3);
+    expect(result).toEqual({ some: 'thing', myParam: '3' });
+  });
+  it('updates a populated query', () => {
+    const query = { some: 'thing', myParam: '1,7,9' };
+    const result = redirectQuery(query, 'myParam', 3);
+    expect(result).toEqual({ some: 'thing', myParam: '1,7,9,3' });
+  });
+});
