+++ ui/components/filters/autocompleter.js
@@ -7,7 +7,8 @@ import React from 'react';
 import { Async } from 'react-select';
 
 import SearchView from './search-view';
-import { apiNameField, redirectQuery, search } from '../lookup-search';
+import { apiNameField, search } from '../../lookup-search';
+import { redirectQuery } from '../../redirects';
 
 export default class Autocompleter extends React.Component {
   constructor(props, context) {
