--- ui/components/redirectWhiteList.js
@@ -1,4 +1,0 @@
-export default [
-  '/policies',
-  '/requirements',
-];
